requirements.txt
Flask==3.0.0
PyPDF2==3.0.1
langchain-text-splitters==0.2.2
langchain-huggingface==0.0.28
langchain-chroma==0.0.28
python-dotenv==1.0.1
anthropic==0.20.0